Moved S3 caching code to its own utils class, fixed Hadoop 2.6.0 build problem
