+++ pom.xml
@@ -1853,6 +1853,10 @@
           <version>${commons-httpclient.version}</version>
           <scope>test</scope>
         </dependency>
+        <dependency>
+          <groupId>org.glassfish.jersey.containers</groupId>
+          <artifactId>jersey-container-servlet-core</artifactId>
+        </dependency>
       </dependencies>
     </profile>
     <profile>
